+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UI,INTERNAL: Provide WithProfileSelection via koin
 - SDK,INTERNAL: Separate Uia logic from UiaRouter
 - SDK,INTERNAL: Separate SendAttachmentRouter from TimelineViewModel
+- SDK,UI (**breaking change**): Sync `ViewState` between SDK and UI. This ensures predictable scroll and loading
+  behavior of the timeline. If the UI does not support setting `ViewState`,
+  `MatrixMessengerConfiguration.timelineViewStateSupported` must be set to false and
+  `TimelienViewModel.finishedScrollTo` additionally be used.
 
 ### Deprecated